Also remap } and {
By default, } and { jump to paragraphs, which is useful in quite a few
filetypes, but in some it's probably more useful to jump to the next
block of code: usually something like if, for, while, etc.

Been using this with Go for a while now, I like it. Need to add more
filetypes.